requirements.txt
PIL
numpy
pandas
pytorch
cudatoolkit=10.0
torchvision
pandas
cython
nb_conda_kernels
nb_conda
ignite
scikit-learn
scikit-plot
easyimages
pretrainedmodels
pytorch-ignite
torchsummary
scikit-image
seaborn
pycocotools
easyimages
matplotlib
sklearn.metrics
skorch 
skorch.callbacks 
skorch.utils
torch
torch.nn
torch.utils.data
torchvision
torch.optim
torch.autograd
torch.optim
os
pickle
datetime
